+++ src/context/item-context.tsx
@@ -5,43 +5,7 @@ import type { ClothingItem, CartItem } from '@/types';
 import { initialItems } from '@/lib/mock-data.ts';
 import { ITEMS_STORAGE_KEY, PURCHASE_COUNTS_STORAGE_KEY } from '@/lib/constants';
 import React, { createContext, useContext, useState, useEffect, useCallback, ReactNode } from 'react';
-
-const ALLOWED_IMAGE_HOSTNAMES = [
-  'placehold.co',
-  'images.unsplash.com',
-  'plus.unsplash.com',
-];
-const DEFAULT_PLACEHOLDER = 'https://placehold.co/600x800.png';
-
-const sanitizeItems = (items: ClothingItem[]): ClothingItem[] => {
-  if (!Array.isArray(items)) return [];
-  return items.map(item => {
-    const imageUrls = Array.isArray(item.imageUrls) ? item.imageUrls : [];
-    const sanitizedUrls = imageUrls.map(url => {
-      if (typeof url !== 'string') return DEFAULT_PLACEHOLDER;
-      try {
-        const hostname = new URL(url).hostname;
-        if (ALLOWED_IMAGE_HOSTNAMES.some(allowedHost => hostname === allowedHost || hostname.endsWith('.' + allowedHost))) {
-          return url;
-        }
-      } catch (e) {
-        // Invalid URL format, fall through to return placeholder
-      }
-      return DEFAULT_PLACEHOLDER;
-    });
-
-    // If after sanitizing, there are no images, add a placeholder
-    if (sanitizedUrls.length === 0) {
-        sanitizedUrls.push(DEFAULT_PLACEHOLDER);
-    }
-
-    return {
-      ...item,
-      imageUrls: sanitizedUrls,
-    };
-  });
-};
-
+import { sanitizeItems } from '@/lib/utils';
 
 interface ItemContextType {
   items: ClothingItem[];
@@ -66,11 +30,11 @@ export const ItemProvider: React.FC<{ children: ReactNode }> = ({ children }) =>
         loadedItems = sanitizeItems(parsedItems);
       } else {
         // If no stored items, use initial (which are assumed to be clean)
-        loadedItems = initialItems;
+        loadedItems = sanitizeItems(initialItems);
       }
     } catch (error) {
       console.warn("Could not access or parse localStorage for items, using initial data:", error);
-      loadedItems = initialItems;
+      loadedItems = sanitizeItems(initialItems);
     } finally {
       setItems(loadedItems);
       // Persist the potentially sanitized data back to localStorage
